Add files via upload
Thanks to a friendly request, I include a simple graph renderer who's output is shown in the README.
Also to note the the "whole module tracing" concept is nice by a bit to powerful. I didn't change that but
arguably, selective tracing is a nice thing, and right now that is not available.